backend/requirements.txt

Django==4.2.5
djangorestframework==3.14.0
python-dotenv==1.0.1
requests==2.31.0
